+++ CHANGELOG.md
@@ -6,12 +6,14 @@ This project adheres to [Semantic Versioning](http://semver.org/). Breaking chan
 This project adheres to the Node [default version scheme](https://docs.npmjs.com/misc/semver).
 
 ## [Next version]
+**Note:** This version might be a breaking change if you are providing customised language translations. Please see [MIGRATION](https://github.com/onfido/onfido-sdk-ui/blob/master/MIGRATION.md).
 
 ### Changed
 - Internal: Redux and EventEmitter are not in the global scope anymore. The `tearDown` function will only unmount the SDK.
 
 ### Added
 - Public: Added a troubleshooting section to the documentation with details about solving CSP related issues
+- UI: Added selfie intro screen
 
 ### Changed
 - UI: Unsupported browser message for mobile browsers without getUserMedia API support when `uploadFallback` option is disabled for live document capture and selfie/liveness capture steps